Extracted bitsize functions to their own file
